requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==4.2.9
django-cors-headers==4.4.0
django-countries==7.6.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
